Move fees module tests to integration
