+++ fold_node/src/fees/lightning/mock.rs
@@ -68,47 +68,3 @@ impl LightningClient for MockLightningClient {
     }
 }
 
-#[cfg(test)]
-mod tests {
-    use super::*;
-
-    #[tokio::test]
-    async fn test_mock_invoice_creation() {
-        let client = MockLightningClient::new();
-        let result = client
-            .create_invoice(
-                100,
-                "Test payment".to_string(),
-                Duration::from_secs(3600),
-                false,
-            )
-            .await;
-
-        assert!(result.is_ok());
-        let invoice = result.unwrap();
-        assert_eq!(invoice.amount, 100);
-        assert!(invoice.invoice.starts_with("mock_invoice_"));
-    }
-
-    #[tokio::test]
-    async fn test_mock_payment_status() {
-        let client = MockLightningClient::new();
-
-        // Test valid mock invoice
-        let status = client.check_payment("mock_invoice_123").await;
-        assert!(status.is_ok());
-        assert!(matches!(status.unwrap(), PaymentStatus::Settled));
-
-        // Test invalid invoice format
-        let status = client.check_payment("invalid_invoice").await;
-        assert!(status.is_err());
-        assert!(matches!(status.unwrap_err(), Error::InvalidInvoice(_)));
-    }
-
-    #[tokio::test]
-    async fn test_mock_node_connection() {
-        let client = MockLightningClient::new();
-        let connected = client.check_node_connection().await.unwrap();
-        assert!(connected);
-    }
-}
