+++ fold_node/src/fees/payment_calculator.rs
@@ -146,137 +146,3 @@ pub fn calculate_total_query_payment(
     Ok(final_payment)
 }
 
-#[cfg(test)]
-mod tests {
-    use super::*;
-    use std::time::Duration;
-
-    fn setup_test_configs() -> (GlobalPaymentConfig, MarketRate, SchemaPaymentConfig) {
-        let global_config = GlobalPaymentConfig::new(
-            50, // 50 sats minimum
-            Duration::from_secs(3600),
-            3,
-            Duration::from_secs(7200),
-        )
-        .unwrap();
-
-        let market_rate = MarketRate::new(100); // 100 sats base rate
-
-        let schema_payment = SchemaPaymentConfig::new(1.5, 10).unwrap();
-
-        (global_config, market_rate, schema_payment)
-    }
-
-    #[test]
-    fn test_calculate_field_payment() {
-        let (global_config, market_rate, schema_payment) = setup_test_configs();
-
-        // Test linear scaling
-        let field_payment = FieldPaymentConfig::new(
-            2.0,
-            TrustDistanceScaling::Linear {
-                slope: 0.5,
-                intercept: 1.0,
-                min_factor: 1.0,
-            },
-            None,
-        )
-        .unwrap();
-
-        let payment = calculate_field_payment(
-            &global_config,
-            &market_rate,
-            &schema_payment,
-            &field_payment,
-            2.0,
-        )
-        .unwrap();
-
-        // Expected: 100 * 1.5 * 2.0 * (0.5 * 2 + 1) = 100 * 1.5 * 2.0 * 2 = 600
-        assert_eq!(payment, 600);
-
-        // Test exponential scaling
-        let field_payment = FieldPaymentConfig::new(
-            2.0,
-            TrustDistanceScaling::Exponential {
-                base: 2.0,
-                scale: 0.5,
-                min_factor: 1.0,
-            },
-            None,
-        )
-        .unwrap();
-
-        let payment = calculate_field_payment(
-            &global_config,
-            &market_rate,
-            &schema_payment,
-            &field_payment,
-            3.0,
-        )
-        .unwrap();
-
-        // Expected: 100 * 1.5 * 2.0 * (2^(0.5 * 3)) ≈ 849
-        assert!(payment > 800 && payment < 900);
-    }
-
-    #[test]
-    fn test_minimum_thresholds() {
-        let (global_config, market_rate, schema_payment) = setup_test_configs();
-
-        // Test field with high minimum payment
-        let field_payment = FieldPaymentConfig::new(
-            0.1, // Very low multiplier
-            TrustDistanceScaling::None,
-            Some(1000), // High minimum threshold
-        )
-        .unwrap();
-
-        let payment = calculate_field_payment(
-            &global_config,
-            &market_rate,
-            &schema_payment,
-            &field_payment,
-            1.0,
-        )
-        .unwrap();
-
-        // Should use field minimum payment
-        assert_eq!(payment, 1000);
-    }
-
-    #[test]
-    fn test_total_query_payment() {
-        let (global_config, market_rate, schema_payment) = setup_test_configs();
-
-        let field1 = FieldPaymentConfig::new(2.0, TrustDistanceScaling::None, None).unwrap();
-
-        let field2 = FieldPaymentConfig::new(1.5, TrustDistanceScaling::None, None).unwrap();
-
-        let fields = vec![(field1, 1.0), (field2, 1.0)];
-
-        let total =
-            calculate_total_query_payment(&global_config, &market_rate, &schema_payment, &fields)
-                .unwrap();
-
-        // Expected: (100 * 1.5 * 2.0) + (100 * 1.5 * 1.5) = 300 + 225 = 525
-        assert_eq!(total, 525);
-    }
-
-    #[test]
-    fn test_negative_trust_distance() {
-        let (global_config, market_rate, schema_payment) = setup_test_configs();
-
-        let field_payment = FieldPaymentConfig::new(2.0, TrustDistanceScaling::None, None).unwrap();
-
-        let result = calculate_field_payment(
-            &global_config,
-            &market_rate,
-            &schema_payment,
-            &field_payment,
-            -1.0,
-        );
-
-        assert!(matches!(result, Err(Error::InvalidTrustDistance(_))));
-    }
-}
