+++ fold_node/src/fees/types/config/mod.rs
@@ -191,61 +191,3 @@ impl Default for FieldPaymentConfig {
     }
 }
 
-#[cfg(test)]
-mod tests {
-    use super::*;
-
-    #[test]
-    fn test_global_payment_config_validation() {
-        // Valid config
-        let config =
-            GlobalPaymentConfig::new(100, Duration::from_secs(3600), 3, Duration::from_secs(7200));
-        assert!(config.is_ok());
-
-        // Invalid system base rate
-        let config =
-            GlobalPaymentConfig::new(0, Duration::from_secs(3600), 3, Duration::from_secs(7200));
-        assert!(config.is_err());
-
-        // Invalid payment timeout
-        let config =
-            GlobalPaymentConfig::new(100, Duration::from_secs(0), 3, Duration::from_secs(7200));
-        assert!(config.is_err());
-
-        // Invalid max retries
-        let config =
-            GlobalPaymentConfig::new(100, Duration::from_secs(3600), 0, Duration::from_secs(7200));
-        assert!(config.is_err());
-    }
-
-    #[test]
-    fn test_schema_payment_config_validation() {
-        // Valid config
-        let config = SchemaPaymentConfig::new(1.5, 10);
-        assert!(config.is_ok());
-
-        // Invalid base multiplier
-        let config = SchemaPaymentConfig::new(0.0, 10);
-        assert!(config.is_err());
-        let config = SchemaPaymentConfig::new(-1.0, 10);
-        assert!(config.is_err());
-    }
-
-    #[test]
-    fn test_market_rate_staleness() {
-        let mut rate = MarketRate::new(100);
-
-        // Not stale immediately
-        assert!(!rate.is_stale(Duration::from_secs(60)));
-
-        // Force last_updated to be in the past
-        rate.last_updated -= chrono::Duration::seconds(120);
-
-        // Should be stale now
-        assert!(rate.is_stale(Duration::from_secs(60)));
-
-        // Update should reset staleness
-        rate.update(150);
-        assert!(!rate.is_stale(Duration::from_secs(60)));
-    }
-}
